move S3Mock to s3 testkit
