+++ s3-test/src/test/scala/geotrellis/spark/io/s3/S3AttributeStoreSpec.scala
@@ -1,0 +1,14 @@
+package geotrellis.spark.io.s3
+
+import geotrellis.spark._
+import geotrellis.spark.io._
+import geotrellis.spark.io.s3.testkit._
+
+class S3AttributeStoreSpec extends AttributeStoreSpec {
+  val bucket = "attribute-store-test-mock-bucket"
+  val prefix = "catalog"
+
+  lazy val attributeStore = new S3AttributeStore(bucket, prefix) {
+    override val s3Client = new MockS3Client()
+  }
+}
