+++ s3-testkit/build.sbt
@@ -1,0 +1,10 @@
+import Dependencies._
+
+name := "geotrellis-s3-testkit"
+libraryDependencies ++= Seq(
+  sparkCore % "provided",
+  awsSdkS3,
+  spire,
+  logging,
+  scalatest
+)
