+++ s3-testkit/src/main/scala/geotrellis/spark/io/s3/testkit/MockS3LayerCollectionReader.scala
@@ -1,0 +1,14 @@
+package geotrellis.spark.io.s3.testkit
+
+import geotrellis.spark.io._
+import geotrellis.spark.io.s3._
+import org.apache.spark._
+
+class MockS3LayerCollectionReader(
+  attributeStore: AttributeStore
+)(implicit sc: SparkContext) extends S3LayerCollectionReader(attributeStore) {
+  override def collectionReader =
+    new S3CollectionReader {
+      def getS3Client = () => new MockS3Client()
+    }
+}
