+++ s3-testkit/src/main/scala/geotrellis/spark/io/s3/testkit/MockS3LayerWriter.scala
@@ -1,0 +1,20 @@
+package geotrellis.spark.io.s3.testkit
+
+import geotrellis.spark.io._
+import geotrellis.spark.io.s3._
+import geotrellis.spark.io.json._
+
+import spray.json._
+
+class MockS3LayerWriter(
+  attributeStore: AttributeStore,
+  bucket: String,
+  keyPrefix: String
+) extends S3LayerWriter(attributeStore, bucket, keyPrefix) {
+  override def rddWriter =
+    new S3RDDWriter {
+      def getS3Client = () => {
+        new MockS3Client()
+      }
+    }
+}
