+++ scripts/buildall.sh
@@ -10,7 +10,9 @@ HOSTALIASES=/tmp/hostaliases ./sbt -J-Xmx2G "project geowave" test || { exit 1; 
 ./sbt -J-Xmx2G "project proj4" test || { exit 1; }
 ./sbt -J-Xmx2G "project raster-test" test || { exit 1; }
 ./sbt -J-Xmx2G "project raster-testkit" compile || { exit 1; }
-./sbt -J-Xmx2G "project s3" test  || { exit 1; }
+./sbt -J-Xmx2G "project s3" compile  || { exit 1; }
+./sbt -J-Xmx2G "project s3-test" test  || { exit 1; }
+./sbt -J-Xmx2G "project s3-testkit" test  || { exit 1; }
 ./sbt -J-Xmx2G "project shapefile" compile || { exit 1; }
 ./sbt -J-Xmx2G "project slick" test:compile || { exit 1; }
 ./sbt -J-Xmx2G "project spark" test  || { exit 1; }
