+++ scripts/cleanall.sh
@@ -7,6 +7,8 @@
 ./sbt -J-Xmx2G "project geowave" clean || { exit 1; }
 ./sbt -J-Xmx2G "project hbase" clean || { exit 1; }
 ./sbt -J-Xmx2G "project proj4" clean || { exit 1; }
+./sbt -J-Xmx2G "project s3" clean || { exit 1; }
+./sbt -J-Xmx2G "project s3-test" clean || { exit 1; }
 ./sbt -J-Xmx2G "project raster-test" clean || { exit 1; }
 ./sbt -J-Xmx2G "project shapefile" clean || { exit 1; }
 ./sbt -J-Xmx2G "project slick" clean || { exit 1; }
@@ -27,6 +29,8 @@ rm -r raster-test/target
 rm -r raster-testkit/target
 rm -r raster/target
 rm -r s3/target
+rm -r s3-test/target
+rm -r s3-testkit/target
 rm -r shapefile/target
 rm -r slick/target
 rm -r spark-testkit/target
